Convert to using external parsing library
Atom and VS Code use the same logic at the parsing level. An Atom
version of this functionality was ported over to using the parsing
library several months ago, and this commit does the same for VS Code's.

The VS Code specific API's and logic are kept in this package, while the
more-generic parsing was moved to the python-indent-parser NPM library.